+++ 30_SoundBlaster16/kernel/ext2.h
@@ -164,7 +164,7 @@ ext2_inode_t * ext2_get_inode(ext2fs_t * fs, uint32_t inode);
 void ext2_read_inode(ext2fs_t * fs, ext2_inode_t * inode, void * buf);
 char * ext2_lsdir(ext2fs_t * fs, ext2_inode_t * inode);
 uint32_t ext2_get_inode_by_name(ext2fs_t * fs, ext2_inode_t * inode, char * name);
-void * ext2_getfile(ext2fs_t * fs, char * path, int);
+void * ext2_getfile(ext2fs_t * fs, char * path, uint16_t);
 void ext2_closefile(void * f);
 size_t ext2_readfile(void * f, void * buf, size_t count);
 void * ext2_readdir(void * f);
